* commit 'e55f302ae5f8f03d095435633a2dd958d9335b74':
  REmoved unused imports
  Same structure in methods in packethelper
  Updated packethelper
  Solved available bug
  Solved bug reading gps data
  Removed udp simple sender

+++ drivers/parrot/src/parrot/ardrone2/util/PacketHelper.java
@@ -33,7 +33,7 @@ public class PacketHelper {
         long value = 0;
         for (int i = 7; i >= 0; i--) {
             int shift = i * 8;
-            value += (data[i + offset] & 0x000000FF) << shift;
+            value += ((long) data[i + offset] & 0x000000FF) << shift;
         }
         return value;
     }
